pyproject.toml
[build-system]
requires = ["setuptools>=42.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fileorganizer"
dynamic = ["version"]
description = "A powerful file organization tool"
readme = "README.md"
requires-python = ">=3.8,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Amber Boudreau", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    # Dependencies will be read from requirements.txt
]

[project.optional-dependencies]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "pytest-mock>=3.0",
    "pytest-xdist>=2.0",
    "coverage>=6.0",
]
dev = [
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
    "mypy>=0.900",
    "pre-commit>=2.0",
    "types-setuptools",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "sphinx-click>=3.0",
    "sphinx-autodoc-typehints>=1.0",
]

[project.urls]
Homepage = "https://github.com/the-solution-desk/fileorganizer"
Documentation = "https://fileorganizer.readthedocs.io/"
Repository = "https://github.com/the-solution-desk/fileorganizer"
Changelog = "https://github.com/the-solution-desk/fileorganizer/blob/main/CHANGELOG.md"

[tool.setuptools.dynamic]
version = {attr = "fileorganizer.__version__"}

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=fileorganizer --cov-report=term-missing"

[tool.coverage.run]
source = ["fileorganizer"]
omit = [
    "*/tests/*",
    "*/__init__.py"
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 90

[flake8]
max-line-length = 88
extend-ignore = "E203, W503"
exclude = ".git,__pycache__,.venv,venv,build,dist"
per-file-ignores = [
    "tests/*: S101",
    "tests/*: B027"
]
